requirements.txt
streamlit
pandas
altair
mysql-connector-python